requirements.txt
Flask==2.0.1
Pillow==8.2.0
transformers==4.9.2
torch==1.9.0
openml==0.12.2
gunicorn==20.1.0
